Set working_directory option in unicorn.config
Set working_directory to the 'current' symlink for the application. This
ensures that the app is upgraded using the current version of the app.
E.g., ensures the Gemfile from the current release of the app is used
when restarting, as opposed to looking back at an old release dir.