Specify Gemfile location in unicorn.config
Specifying the Gemfile location ensures unicorn will use the Gemfile
from the current deployment directory. Otherwise, unicorn will continue
to load the Gemfile from the same directory as when the first master
process was first started until it is explicitly stopped and then
started again.

This is the solution recommended by unicorn:
http://unicorn.bogomips.org/Sandbox.html